requirements.txt
PIL>=1.1.6
Django>=1.4
boto>=2.5
django-storages>=1.1.8